refactor: add evaluators module with cross-validation and hold-out validation functions; update documentation and tests
