--- docs/api/utils/evaluators.md
@@ -1,1 +1,0 @@
-::: siapy.utils.evaluators
