--- siapy/utils/enums.py
@@ -1,11 +1,0 @@
-from enum import Enum, auto
-
-__all__ = [
-    "InteractiveButtonsEnum",
-]
-
-
-class InteractiveButtonsEnum(Enum):
-    SAVE = auto()
-    REPEAT = auto()
-    SKIP = auto()
